app/requirements.txt
cassandra-driver==3.18.0
Flask==1.1.2
numpy==1.19.1
Keras==2.3.1
Pillow==6.2.1
tensorflow==2.1.0
matplotlib==3.2.2

